Move config files to new gui lib
