+++ src/gui/src/MainWindow.h
@@ -29,13 +29,13 @@
 #include "ui_MainWindowBase.h"
 
 #include "ActivationDialog.h"
-#include "AppConfig.h"
 #include "ClientConnection.h"
-#include "Config.h"
 #include "ServerConfig.h"
 #include "ServerConnection.h"
 #include "TrayIcon.h"
 #include "global/Ipc.h"
+#include "gui/AppConfig.h"
+#include "gui/Config.h"
 #include "gui/QIpcClient.h"
 #include "gui/TlsUtility.h"
 #include "gui/VersionChecker.h"
