+++ src/gui/src/ServerConfigDialog.h
@@ -20,10 +20,9 @@
 
 #define SERVERCONFIGDIALOG__H
 
-#include "AppConfig.h"
 #include "ScreenSetupModel.h"
 #include "ServerConfig.h"
-
+#include "gui/AppConfig.h"
 #include "ui_ServerConfigDialogBase.h"
 
 #include <QDialog>
