+++ src/lib/gui/AppConfig.cpp
